refactor（v0.2.1-optimal）：拆分 protocol 根转换逻辑
