+++ src-tauri/src/protocol/legacy/mod.rs
@@ -1,0 +1,12 @@
+mod anthropic_decode;
+mod anthropic_encode;
+mod responses_decode;
+mod responses_encode;
+
+#[cfg(test)]
+mod tests;
+
+pub use anthropic_decode::{anthropic_to_openai, estimate_anthropic_input_tokens};
+pub use anthropic_encode::openai_to_anthropic;
+pub use responses_decode::responses_to_openai;
+pub use responses_encode::openai_to_responses;
