+++ src-tauri/src/protocol/legacy/responses_encode.rs
@@ -1,0 +1,92 @@
+use serde_json::Value;
+
+/// Convert OpenAI Chat Completions response to Responses API format.
+pub fn openai_to_responses(openai_resp: &Value, model: &str) -> Value {
+    let choice = openai_resp
+        .get("choices")
+        .and_then(|c| c.as_array())
+        .and_then(|arr| arr.first());
+
+    let message = choice.and_then(|ch| ch.get("message"));
+
+    let content = message
+        .and_then(|msg| msg.get("content"))
+        .and_then(|c| c.as_str())
+        .unwrap_or("");
+
+    let finish_reason = choice
+        .and_then(|ch| ch.get("finish_reason"))
+        .and_then(|f| f.as_str())
+        .unwrap_or("stop");
+
+    let prompt_tokens = openai_resp
+        .get("usage")
+        .and_then(|u| u.get("prompt_tokens"))
+        .and_then(|t| t.as_u64())
+        .unwrap_or(0);
+    let completion_tokens = openai_resp
+        .get("usage")
+        .and_then(|u| u.get("completion_tokens"))
+        .and_then(|t| t.as_u64())
+        .unwrap_or(0);
+
+    // Build output array: message + function_call items
+    let mut output = Vec::new();
+
+    // Add function_call outputs for tool_calls
+    if let Some(tool_calls) = message
+        .and_then(|m| m.get("tool_calls"))
+        .and_then(|t| t.as_array())
+    {
+        for tc in tool_calls {
+            let name = tc
+                .get("function")
+                .and_then(|f| f.get("name"))
+                .and_then(|n| n.as_str())
+                .unwrap_or("");
+            let arguments = tc
+                .get("function")
+                .and_then(|f| f.get("arguments"))
+                .and_then(|a| a.as_str())
+                .unwrap_or("");
+            let call_id = tc.get("id").and_then(|i| i.as_str()).unwrap_or("");
+            output.push(serde_json::json!({
+                "id": format!("fc_{}", uuid::Uuid::new_v4().simple()),
+                "type": "function_call",
+                "call_id": call_id,
+                "name": name,
+                "arguments": arguments,
+                "status": "completed"
+            }));
+        }
+    }
+
+    // Add text message output (always include, even if empty when tool_calls present)
+    if !content.is_empty() || output.is_empty() {
+        output.push(serde_json::json!({
+            "id": format!("msg_{}", uuid::Uuid::new_v4().simple()),
+            "type": "message",
+            "role": "assistant",
+            "content": [{
+                "type": "output_text",
+                "text": content
+            }],
+            "status": "completed"
+        }));
+    }
+
+    serde_json::json!({
+        "id": openai_resp.get("id").cloned().unwrap_or(Value::String(format!("resp_{}", uuid::Uuid::new_v4()))),
+        "object": "response",
+        "created_at": chrono::Utc::now().timestamp(),
+        "model": model,
+        "output": output,
+        "usage": {
+            "input_tokens": prompt_tokens,
+            "output_tokens": completion_tokens,
+            "total_tokens": prompt_tokens + completion_tokens
+        },
+        "status": "completed",
+        "finish_reason": finish_reason
+    })
+}
