+++ src-tauri/src/protocol/legacy/tests/mod.rs
@@ -1,0 +1,2 @@
+mod anthropic;
+mod responses;
